Initial Commit - Apache-Highlighter
Syntax Highlight Python and Send it as HTML

Requires Apache with mod_cgi and mod_actions
  and Python with the Pygments module installed

To install, place it in your /cgi-bin directory and make sure it's
readable and executable.

Add the following lines into the main Apache config, a directive block in
your Apache config, or into an .htaccess file in individual directories.

Add to or remove from this list to your liking. Only the first line and
one AddHandler line are necessary.

    Action highlight /cgi-bin/highlight.cgi
    AddHandler highlight .py
    AddHandler highlight .pl
    AddHandler highlight .pm
    AddHandler highlight .perl
    AddHandler highlight .s
    AddHandler highlight .c
    AddHandler highlight .sh
    AddHandler highlight .html
    AddHandler highlight .cpp
    AddHandler highlight .xml
    AddHandler highlight .js
    AddHandler highlight .css

The full list of highlightable syntaxes is available at
http://pygments.org/languages/

WARNING: Be very sure that the directory to which you apply this contains
the correct files you want to highlight. This can be a major security flaw
if you share scripts and other files you don't intend to.
Signed-off-by: Jonas Buckner <buckner.jonas@gmail.com>